add source command
This source command was stolen from the bot repo then adapted to work here. Little was changed. There is room for refactor, but this PR is relatively urgent, would prefer to refactor later. This functionality should probably be moved to BotCore ultimately, so a total refactor may not be worth the time.