Introduce `pin flow` to the Code Jam Extension
To avoid code duplication an additional flow was introduced, the `pin_flow`.

Additionally, the `message_is_in_code_jam_category` was deleted, and an extra check was integrated into the `pin_flow` to avoid permission problems, by checking that the referenced message's channel is the same as the context's channel.