Add functionality for a true silent fail
This commit makes a compromise. The top level command group
can be run anywhere.
This allows the sub-group decorators to behave properly.
Now, if a `role_override` is provided to the
`in_whitelist` check, anyone with that role
does not have to obey channel restrictions.

This commits also adds a more explicit SilentFail error to catch.
If you want something to truly silently fail, we can now do that.